Updated links in Readme.md
Removed 3rd link, "nuxt 3 support on Azure Static Web Apps," since I didn't know where that was supposed to point.  Nuxt 3 support is discussed in the second link, to learn.microsoft.com, so it seems redundant, plus it was just pointing to self (#).  Fixed the link to deployment documentation at the end of the file, which was pointing to an older version.